requirements.txt
amqp==5.2.0
asgiref==3.8.1
atatus==1.5.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==4.2.6
django-celery==3.1.17
django-cors-headers==4.4.0
django-extensions==3.2.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
git-filter-repo==2.45.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
mysqlclient==2.2.4
pillow==10.4.0
prompt_toolkit==3.0.45
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.32.3
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zope.interface==6.4.post2
